move sanitize_array to core.construction
